# config/openapi/wrk-aircraft-api.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Aircraft REST CRUD operations API in Spring-Boot 2",
        "description": "REST API for monitoring using Spring Boot, Prometheus and Graphana ",
        "termsOfService": "miclam",
        "contact": {
            "name": "Michel Lamrani",
            "url": "https://miclam",
            "email": "https://miclam"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://wrk-aircraft.dev.k8s-qa.dassault-avion.fr:80",
            "description": "Inferred Url"
        }
    ],
...
    "paths": {
        "/aircrafts": {
            "get": {
                "tags": [
                    "aircraft-controller"
                ],
                "summary": "getAllAircraft",
                "operationId": "getAllAircraftUsingGET",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Aircraft"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "aircraft-controller"
                ],
                "summary": "updateAircraft",
                "operationId": "updateAircraftUsingPUT",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Aircraft"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Aircraft"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "aircraft-controller"
                ],
                "summary": "addNewAircraft",
                "operationId": "addNewAircraftUsingPOST",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Aircraft"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Aircraft"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/aircrafts/{aircraftId}": {
            "get": {
                "tags": [
                    "aircraft-controller"
                ],
                "summary": "getAircraft",
                "operationId": "getAircraftUsingGET",
                "parameters": [
                    {
                        "name": "aircraftId",
                        "in": "path",
                        "description": "aircraftId",
                        "required": true,
                        "style": "simple",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Aircraft"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "aircraft-controller"
                ],
                "summary": "deleteAircraft",
                "operationId": "deleteAircraftUsingDELETE",
                "parameters": [
                    {
                        "name": "aircraftId",
                        "in": "path",
                        "description": "aircraftId",
                        "required": true,
                        "style": "simple",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/app/env": {
            "get": {
                "tags": [
                    "app-controller"
                ],
                "summary": "getAppEnv",
                "operationId": "getAppEnvUsingGET",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/app/ip": {
            "get": {
                "tags": [
                    "app-controller"
                ],
                "summary": "getAppIP",
                "operationId": "getAppIPUsingGET",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/app/name": {
            "get": {
                "tags": [
                    "app-controller"
                ],
                "summary": "getAppName",
                "operationId": "getAppNameUsingGET",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/load/cpu": {
            "get": {
                "tags": [
                    "load-controller"
                ],
                "summary": "loadCPU",
                "operationId": "loadCPUUsingGET",
                "parameters": [
                    {
                        "name": "core",
                        "in": "query",
                        "description": "core",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "th",
                        "in": "query",
                        "description": "th",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "du",
                        "in": "query",
                        "description": "du",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/load/ram": {
            "get": {
                "tags": [
                    "load-controller"
                ],
                "summary": "loadMemory",
                "operationId": "loadMemoryUsingGET",
                "parameters": [
                    {
                        "name": "mem",
                        "in": "query",
                        "description": "mem",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "wait",
                        "in": "query",
                        "description": "wait",
                        "required": false,
                        "style": "form",
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Aircraft": {
                "title": "Aircraft",
                "type": "object",
                "properties": {
                    "aircraftId": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "category": {
                        "type": "string"
                    },
                    "model": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}